Update: Added: Refactoring: Remove value sid from report and print
