+++ .github/workflows/test_api.yml
@@ -32,8 +32,7 @@ jobs:
           EMAIL: ${{ secrets.EMAIL}}
           PASSWORD: ${{ secrets.PASSWORD}}
         run: |
-          pytest -s --alluredir=allure-results 
-
+          pytest -s --alluredir=allure-results
       - name: Add allure information
         working-directory: allure-results
         if: always()
