+++ tests_ui/test_login_form.py
@@ -1,0 +1,3 @@
+def test_fo():
+    print('*' * 50, '*' * 50)
+    assert 1 == 1
